requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.2.8
django-crispy-forms==1.13.0
django-environ==0.9.0
django-heroku==0.0.0
django-jazzmin==2.5.0
django-resized==0.3.11
django-summernote==0.8.20.0
django-taggit==2.1.0
gunicorn==20.1.0
idna==3.3
lxml==4.8.0
markdown2==2.4.2
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.3
pyparsing==3.0.7
pyquery==1.4.3
pytz==2021.3
readtime==1.1.1
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
uuid==1.30
webencodings==0.5.1
whitenoise==6.0.0
